Loader() should usually be stored with the page it will be used in  -making route definitions leaner
