+++ frontend/src/Pages/Events.jsx
@@ -1,11 +1,22 @@
-// import { useLoaderData } from 'react-router-dom';
+import { useLoaderData } from 'react-router-dom';
 import EventsList from '../components/EventsList';
 
 function EventsPage() {
 
-  // const events = useLoaderData();
+  const events = useLoaderData();
 
-  return <EventsList  />;
+  return <EventsList events={events} />;
 }
 
 export default EventsPage;
+
+export async function loader() {
+  const response = await fetch('http://localhost:8080/events');
+
+  if (!response.ok) {
+    //...
+  } else {
+    const resData = await response.json();
+    return resData.events;
+  }
+}
